requirements.txt
bcrypt==4.0.1
capstone==5.0.0rc2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colored-traceback==0.3.0
cryptography==38.0.4
gmpy2==2.1.2
idna==3.4
intervaltree==3.1.0
Mako==1.2.4
MarkupSafe==2.1.1
packaging==21.3
paramiko==2.12.0
pathlib2==2.3.7.post1
pkg_resources==0.0.0
plumbum==1.8.0
psutil==5.9.4
pwntools==4.8.0
pycparser==2.21
pycryptodome==3.16.0
pyelftools==0.29
Pygments==2.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.28.1
ROPGadget==7.1
rpyc==5.3.0
six==1.16.0
sortedcontainers==2.4.0
unicorn==2.0.1.post1
urllib3==1.26.13
